revert: restore PULL_REQUEST_TEMPLATE.md and fix build
